pyproject.toml
[tool.poetry]
name = "smart-blog"
version = "0.1.0"
description = ""
authors = ["Pedriq <1pedre.off@gmail.com>"]
readme = "README.md"
packages = [{include = "smart_blog"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
pillow = "^9.5.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
